--- crates/evm-prover/src/prover/programs/types.rs
@@ -1,18 +1,0 @@
-use alloy_provider::fillers::FillProvider;
-
-pub type DefaultProvider = FillProvider<
-    alloy_provider::fillers::JoinFill<
-        alloy_provider::Identity,
-        alloy_provider::fillers::JoinFill<
-            alloy_provider::fillers::GasFiller,
-            alloy_provider::fillers::JoinFill<
-                alloy_provider::fillers::BlobGasFiller,
-                alloy_provider::fillers::JoinFill<
-                    alloy_provider::fillers::NonceFiller,
-                    alloy_provider::fillers::ChainIdFiller,
-                >,
-            >,
-        >,
-    >,
-    alloy_provider::RootProvider,
->;
