+++ crates/evm-prover/tests/integration.rs
@@ -5,11 +5,8 @@ use std::{
 
 use alloy_primitives::Address;
 use alloy_provider::ProviderBuilder;
-use evm_prover::prover::programs::{
-    message::{AppContext, HyperlaneMessageProver, MerkleTreeState},
-    types::DefaultProvider,
-    MockStateQueryProvider,
-};
+use evm_prover::prover::programs::message::{AppContext, HyperlaneMessageProver, MerkleTreeState};
+use evm_state_queries::hyperlane::{DefaultProvider, MockStateQueryProvider};
 use evm_storage_proofs::client::EvmClient;
 use reqwest::Url;
 use storage::hyperlane::{message::HyperlaneMessageStore, snapshot::HyperlaneSnapshotStore};
